Modernized the construction of the CVE-format
After I went and looked through the code I had the urge to rewrite that
part. It should be behave the same but I was not able to test it
properly.

In addition the construction of the description deserves an f-strings.

Again I was not able to test this properly, without mocking
code I don't fully grasp.